mender-setup.bbclass: stop globally exporting MENDER_MACHINE
This patch is ported from an unmerged change to meta-mender, but
made instead fixed up here in meta-testdistro for convenience.

The original comments:

mender-setup.bbclass: stop globally exporting MENDER_MACHINE
and the accompanying change to BB_HASHBASE_WHITELIST which
unsuccesfully tries to prevent signature changes to *every*
*single* *recipe* being built.

The only reason this variable is there is to satisfy some
test code, while making it impossible to have common
shared-state artifacts for packages between builds that
include this layer and builds that don't.  Very rude.